fix: fixed expression generation where right shift generated left shift, and vice versa, the same with OR and AND operations
